Update deprecated function calls, addition of some comments.

+++ code/computeDifference.py
@@ -2,10 +2,8 @@
 
 def computeDifference(array1,array2,type):
     if type == 'D':
-        difference=array1!=array2
-        difference = difference.astype(float)
+        difference = (array1!=array2).astype(float)
     if type == 'C':
-        difference=abs(array1-array2)
-
+        difference = abs(array1 - array2)
 
     return difference
